Add Grumble 1.0.
This is the first release of Grumble, my print-debugging library for
Python that helps keep console log output simple while also providing
easily-found details in dedicated log files.

It's useful when in the depths of a debugging session, where state
matters and repeated runs need to be scrutinized.

There might be some other things going on in the module too. Who can
say.

See the GitHub repository's documentation for details.